Initial commit
Created manpages as mandated by https:/xkcd.com/1692
Since the section of the main man page was not specified, it is in section 6
for now (blerp.6). The other pages have been stubbed out but should eventually
be worth reading in their own right.

Consider adding blurb(???) page for completeness, but that's more of a stretch
goal.